add yaml and events tab to blockpool
