+++ packages/ocs/block-pool/BlockPoolDetailsPage.tsx
@@ -4,6 +4,7 @@ import { LoadingBox } from '@odf/shared/generic/status-box';
 import { Kebab } from '@odf/shared/kebab/kebab';
 import { useModalLauncher } from '@odf/shared/modals/modalLauncher';
 import { referenceForModel } from '@odf/shared/utils';
+import { EventStreamWrapped, YAMLEditorWrapped } from '@odf/shared/utils/Tabs';
 import { useK8sWatchResource } from '@openshift-console/dynamic-plugin-sdk';
 import { useTranslation } from 'react-i18next';
 import { RouteComponentProps, useLocation } from 'react-router-dom';
@@ -86,6 +87,16 @@ export const BlockPoolDetailsPage: React.FC<BlockPoolDetailsPageProps> = ({
               href: '',
               name: 'Details',
               component: BlockPoolDashboard as any,
+            },
+            {
+              href: 'yaml',
+              name: 'YAML',
+              component: YAMLEditorWrapped,
+            },
+            {
+              href: 'events',
+              name: 'Events',
+              component: EventStreamWrapped,
             },
           ]}
         />
